@@ -1,7 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <ProjectGuid>34F97D69-6382-45C6-A470-2F4065C8FB33</ProjectGuid>
+    <LangVersion>latest</LangVersion>
   </PropertyGroup>
   <ItemGroup>
     <None Remove="Dockerfile" />
@@ -15,6 +16,6 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="IkeMtz.NRSRx.Core.SignalR" Version="1.20.171.5" />
+    <PackageReference Include="IkeMtz.NRSRx.Core.SignalR" Version="6.0.184.6" />
   </ItemGroup>
 </Project>
