@@ -1,13 +1,12 @@
-using NurseCron.Certifications.OData;
-using NurseCron.Certifications.OData.Configuration;
 using IkeMtz.NRSRx.Core.Unigration;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.Extensions.Configuration;
+using NurseCron.Certifications.OData;
 
 namespace NurseCron.Certifications.Tests.Integration
 {
   public class IntegrationODataTestStartup
-      : CoreODataIntegrationTestStartup<Startup, CertificationConfiguration>
+      : CoreODataIntegrationTestStartup<Startup>
   {
     public IntegrationODataTestStartup(IConfiguration configuration)
         : base(new Startup(configuration))
