@@ -1,14 +1,13 @@
-using NurseCron.Certifications.OData;
-using NurseCron.Certifications.OData.Configuration;
-using NurseCron.Certifications.OData.Data;
 using IkeMtz.NRSRx.Core.Unigration;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using NurseCron.Certifications.OData;
+using NurseCron.Certifications.OData.Data;
 
 namespace NurseCron.Certifications.Tests
 {
   public class UnigrationODataTestStartup
-      : CoreODataUnigrationTestStartup<Startup, CertificationConfiguration>
+      : CoreODataUnigrationTestStartup<Startup>
   {
     public UnigrationODataTestStartup(IConfiguration configuration) : base(new Startup(configuration))
     {
