@@ -1,11 +1,12 @@
 using System.Diagnostics.CodeAnalysis;
 using System.Reflection;
-using NurseCron.Competencies.Abstraction.Models;
-using NurseCron.Competencies.OData.Data;
 using IkeMtz.NRSRx.Core.OData;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using NurseCron.Competencies.Abstraction.Models;
+using NurseCron.Competencies.OData.Configuration;
+using NurseCron.Competencies.OData.Data;
 
 namespace NurseCron.Competencies.OData
 {
@@ -16,18 +17,14 @@ namespace NurseCron.Competencies.OData
     public Startup(IConfiguration configuration) : base(configuration)
     {
     }
+    public override BaseODataModelProvider ODataModelProvider => new ODataModelProvider();
 
     [ExcludeFromCodeCoverage]
     public override void SetupDatabase(IServiceCollection services, string connectionString)
     {
-      services
-      .AddDbContextPool<CompetenciesContext>(x => x.UseSqlServer(connectionString))
-      .AddEntityFrameworkSqlServer();
-    }
-
-    public override void SetupMiscDependencies(IServiceCollection services)
-    {
-      services.AddScoped<ICompetenciesContext, CompetenciesContext>();
+      _ = services
+       .AddDbContextPool<DatabaseContext>(x => x.UseSqlServer(connectionString))
+       .AddEntityFrameworkSqlServer();
     }
   }
 }
