@@ -1,13 +1,12 @@
-using NurseCron.Competencies.OData;
-using NurseCron.Competencies.OData.Configuration;
 using IkeMtz.NRSRx.Core.Unigration;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.Extensions.Configuration;
+using NurseCron.Competencies.OData;
 
 namespace NurseCron.Competencies.Tests.Integration
 {
   public class IntegrationTestStartup
-      : CoreODataIntegrationTestStartup<Startup, CompetencyConfiguration>
+      : CoreODataIntegrationTestStartup<Startup>
   {
     public IntegrationTestStartup(IConfiguration configuration)
         : base(new Startup(configuration))
