@@ -3,7 +3,7 @@ using NurseCron.Competencies.WebApi;
 using NurseCron.Competencies.WebApi.Data;
 using IkeMtz.NRSRx.Core.Unigration;
 using IkeMtz.NRSRx.Core.Unigration.WebApi;
-using Microsoft.Azure.ServiceBus;
+using Azure.Messaging.ServiceBus;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -17,12 +17,12 @@ namespace NurseCron.Competencies.Tests.Unigration
 
     public override void SetupDatabase(IServiceCollection services, string connectionString)
     {
-      services.SetupTestDbContext<CompetenciesContext>();
+      services.SetupTestDbContext<DatabaseContext>();
     }
 
     public override void SetupPublishers(IServiceCollection services)
     {
-      var pubTester = new PublisherUnigrationTester<Competency, Message>();
+      var pubTester = new PublisherUnigrationTester<Competency, ServiceBusMessage>();
       pubTester.RegisterDependencies(services);
     }
   }
