@@ -1,38 +1,33 @@
 using System.Linq;
 using IkeMtz.NRSRx.Core.Models;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.OData.Query;
+using Microsoft.AspNetCore.OData.Routing.Controllers;
+using Microsoft.EntityFrameworkCore;
 using NurseCron.Employees.Models;
 using NurseCron.Employees.OData.Data;
-using Microsoft.AspNet.OData;
-using Microsoft.AspNet.OData.Routing;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-using static Microsoft.AspNet.OData.Query.AllowedQueryOptions;
 using static Microsoft.AspNetCore.Http.StatusCodes;
 
 namespace NurseCron.Employees.OData.Controllers.V1
 {
-  //TODO: Need to re-add Authorize attribute
+  [Authorize]
   [ApiVersion(VersionDefinitions.V1_0)]
-  [ODataRoutePrefix(nameof(Employees))]
   public class EmployeesController : ODataController
   {
-    private readonly IEmployeesContext _ctx;
+    private readonly DatabaseContext _ctx;
 
-    public EmployeesController(IEmployeesContext ctx)
+    public EmployeesController(DatabaseContext ctx)
     {
       _ctx = ctx;
     }
 
     [Produces("application/json")]
-    [ODataRoute]
     [ProducesResponseType(typeof(ODataEnvelope<Employee>), Status200OK)]
-    [EnableQuery(MaxTop = 100, AllowedQueryOptions = All)]
+    [EnableQuery(MaxTop = 100, AllowedQueryOptions = AllowedQueryOptions.All)]
     public IQueryable Get()
     {
       return _ctx.Employees
-        .Include(t => t.EmployeeCertifications)
-        .Include(t => t.EmployeeCompetencies)
-        .Include(t => t.EmployeeHealthItems)
         .Where(t => t.IsEnabled);
     }
   }
