@@ -3,12 +3,12 @@ using System.Linq;
 using System.Threading.Tasks;
 using IkeMtz.NRSRx.Core.Models;
 using IkeMtz.NRSRx.Core.Unigration;
+using Microsoft.AspNetCore.TestHost;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using NurseCron.Employees.Models;
 using NurseCron.Employees.OData;
 using NurseCron.Employees.OData.Data;
-using Microsoft.AspNetCore.TestHost;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Newtonsoft.Json;
+using static IkeMtz.NRSRx.Core.Unigration.TestDataFactory;
 
 namespace NurseCron.Employees.Tests.Unigration.OData
 {
@@ -19,32 +19,32 @@ namespace NurseCron.Employees.Tests.Unigration.OData
     [TestCategory("Unigration")]
     public async Task GetEnabledEmployeesTest()
     {
-      var objA = new Employee()
-      {
-        Id = Guid.NewGuid(),
-        FirstName = "Test",
-        IsEnabled = true,
-        BirthDate = DateTime.Today.AddYears(-20),
-        CreatedOnUtc = DateTime.UtcNow,
-      };
+      var objA = CreateIdentifiable(CreateAuditable<Employee>());
+
+      objA.FirstName = "Test";
+      objA.IsEnabled = true;
+      objA.BirthDate = DateTime.Today.AddYears(-20);
+      objA.Email = "me@me.com";
+      objA.LastName = StringGenerator(20);
+
       using var srv = new TestServer(TestHostBuilder<Startup, UnigrationODataTestStartup>()
           .ConfigureTestServices(x =>
             {
-              ExecuteOnContext<EmployeesContext>(x, db =>
+              ExecuteOnContext<DatabaseContext>(x, db =>
                     {
-                      db.Employees.Add(objA);
+                      _ = db.Employees.Add(objA);
                     });
             })
        );
       var client = srv.CreateClient();
       GenerateAuthHeader(client, GenerateTestToken());
 
-      var resp = await client.GetStringAsync("odata/v1/Employees?$count=true");
-
-      TestContext.WriteLine($"Server Reponse: {resp}");
-      var envelope = JsonConvert.DeserializeObject<ODataEnvelope<Employee>>(resp);
-      Assert.AreEqual(objA.CreatedOnUtc, envelope.Value.First().CreatedOnUtc.ToUniversalTime());
+      var resp = await client.GetAsync($"odata/v1/{nameof(Employee)}s?$count=true");
+       
+      var envelope = await DeserializeResponseAsync<Employee[]>(resp);
+      Assert.AreEqual(objA.CreatedOnUtc, envelope.First().CreatedOnUtc.ToUniversalTime());
     }
+
     [TestMethod]
     [TestCategory("Unigration")]
     public async Task HideDisabledEmployeesTest()
@@ -53,25 +53,27 @@ namespace NurseCron.Employees.Tests.Unigration.OData
       {
         Id = Guid.NewGuid(),
         FirstName = "Test",
-        IsEnabled = false
+        IsEnabled = false,
+        Email = "me@me.com",
+        LastName = StringGenerator(20),
       };
       using var srv = new TestServer(TestHostBuilder<Startup, UnigrationODataTestStartup>()
           .ConfigureTestServices(x =>
           {
-            ExecuteOnContext<EmployeesContext>(x, db =>
+            ExecuteOnContext<DatabaseContext>(x, db =>
                   {
-                    db.Employees.Add(objA);
+                    _ = db.Employees.Add(objA);
                   });
           })
        );
       var client = srv.CreateClient();
       GenerateAuthHeader(client, GenerateTestToken());
 
-      var resp = await client.GetAsync("odata/v1/Employees?$count=true");
+      var resp = await client.GetAsync($"odata/v1/{nameof(Employee)}s?$count=true");
 
-      var objB = await DeserializeResponseAsync<ODataEnvelope<Employee>>(resp);
+      var objB = await DeserializeResponseAsync<Employee[]>(resp);
 
-      Assert.AreEqual(0, objB.Value.Count());
+      Assert.AreEqual(0, objB.Length);
     }
 
   }
