@@ -2,14 +2,15 @@ using System;
 using System.Net;
 using System.Threading.Tasks;
 using IkeMtz.NRSRx.Core.Unigration;
+using IkeMtz.NRSRx.Core.Unigration.Http;
 using IkeMtz.NRSRx.Core.WebApi;
-using NurseCron.HealthItems.Models;
-using NurseCron.HealthItems.WebApi;
-using NurseCron.HealthItems.WebApi.Data;
 using Microsoft.AspNetCore.TestHost;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Newtonsoft.Json;
+using NurseCron.HealthItems.Models;
+using NurseCron.HealthItems.WebApi;
+using NurseCron.HealthItems.WebApi.Data;
 
 namespace NurseCron.HealthItems.Tests.Unigration.Api
 {
@@ -51,7 +52,7 @@ namespace NurseCron.HealthItems.Tests.Unigration.Api
 
       Assert.AreEqual("IntegrationTester@email.com", result.CreatedBy);
 
-      var ctx = srv.GetDbContext<HealthItemsContext>();
+      var ctx = srv.GetDbContext<DatabaseContext>();
       var item = await ctx.HealthItems.FirstOrDefaultAsync(t => t.Name == result.Name);
 
       Assert.IsNotNull(item);
@@ -71,7 +72,7 @@ namespace NurseCron.HealthItems.Tests.Unigration.Api
       using var srv = new TestServer(TestHostBuilder<Startup, UnigrationWebApiTestStartup>()
         .ConfigureTestServices(x =>
         {
-          ExecuteOnContext<HealthItemsContext>(x, db =>
+          ExecuteOnContext<DatabaseContext>(x, db =>
           {
             _ = db.HealthItems.Add(objA);
           });
@@ -87,7 +88,7 @@ namespace NurseCron.HealthItems.Tests.Unigration.Api
       Assert.AreEqual("IntegrationTester@email.com", result.UpdatedBy);
       Assert.AreEqual(objA.Name, result.Name);
 
-      var ctx = srv.GetDbContext<HealthItemsContext>();
+      var ctx = srv.GetDbContext<DatabaseContext>();
       var item = await ctx.HealthItems.FirstOrDefaultAsync(t => t.Name == result.Name);
 
       Assert.IsNotNull(item);
@@ -126,7 +127,7 @@ namespace NurseCron.HealthItems.Tests.Unigration.Api
       using var srv = new TestServer(TestHostBuilder<Startup, UnigrationWebApiTestStartup>()
         .ConfigureTestServices(x =>
         {
-          ExecuteOnContext<HealthItemsContext>(x, db =>
+          ExecuteOnContext<DatabaseContext>(x, db =>
           {
             _ = db.HealthItems.Add(objA);
           });
