@@ -3,7 +3,7 @@ using IkeMtz.NRSRx.Core.Unigration.WebApi;
 using NurseCron.HealthItems.Models;
 using NurseCron.HealthItems.WebApi;
 using NurseCron.HealthItems.WebApi.Data;
-using Microsoft.Azure.ServiceBus;
+using Azure.Messaging.ServiceBus;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -17,12 +17,12 @@ namespace NurseCron.HealthItems.Tests.Unigration
 
     public override void SetupDatabase(IServiceCollection services, string connectionString)
     {
-      services.SetupTestDbContext<HealthItemsContext>();
+      services.SetupTestDbContext<DatabaseContext>();
     }
 
     public override void SetupPublishers(IServiceCollection services)
     {
-      new PublisherUnigrationTester<HealthItem, Message>().RegisterDependencies(services);
+      new PublisherUnigrationTester<HealthItem, ServiceBusMessage>().RegisterDependencies(services);
     }
   }
 }
