@@ -1,12 +1,11 @@
 using IkeMtz.NRSRx.Core.Unigration;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.Extensions.Configuration;
-using NurseCron.Services.Schedules.OData.Configuration;
 
 namespace NurseCron.Services.Schedules.OData.Tests.Integration
 {
   public class IntegrationODataTestStartup
-      : CoreODataIntegrationTestStartup<Startup, ModelConfiguration>
+      : CoreODataIntegrationTestStartup<Startup>
   {
     public IntegrationODataTestStartup(IConfiguration configuration)
         : base(new Startup(configuration))
