@@ -1,13 +1,12 @@
 using IkeMtz.NRSRx.Core.Unigration;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using NurseCron.Services.Schedules.OData.Configuration;
 using NurseCron.Services.Schedules.OData.Data;
 
 namespace NurseCron.Services.Schedules.OData.Tests.Unigration
 {
   public class UnigrationODataTestStartup
-      : CoreODataUnigrationTestStartup<Startup, ModelConfiguration>
+      : CoreODataUnigrationTestStartup<Startup>
   {
     public UnigrationODataTestStartup(IConfiguration configuration) : base(new Startup(configuration))
     {
