@@ -3,7 +3,7 @@ using Microsoft.EntityFrameworkCore;
 
 namespace NurseCron.Units.OData.Data
 {
-  public class DatabaseContext : DbContext, IDatabaseContext
+  public class DatabaseContext : DbContext
   {
     public DatabaseContext(DbContextOptions<DatabaseContext> options)
         : base(options)
