@@ -1,14 +1,13 @@
 using IkeMtz.NRSRx.Core.Unigration;
-using NurseCron.Units.OData;
-using NurseCron.Units.OData.Configuration;
-using NurseCron.Units.OData.Data;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using NurseCron.Units.OData;
+using NurseCron.Units.OData.Data;
 
 namespace NurseCron.Units.Tests.Unigration.OData
 {
   public class UnigrationODataTestStartup
-      : CoreODataUnigrationTestStartup<Startup, ModelConfiguration>
+      : CoreODataUnigrationTestStartup<Startup>
   {
     public UnigrationODataTestStartup(IConfiguration configuration) : base(new Startup(configuration))
     {
